% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_counts}
\alias{cell_counts}
\title{Example data set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 720 rows and 5 columns.
}
\usage{
cell_counts
}
\description{
Example data set
}
\keyword{datasets}
