% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_and_install_cmdstanr}
\alias{check_and_install_cmdstanr}
\title{Check and Install cmdstanr and CmdStan}
\usage{
check_and_install_cmdstanr()
}
\description{
This function checks if the \code{cmdstanr} package and CmdStan are installed.
If they are not installed, it installs them automatically in non-interactive sessions
or asks for permission to install them in interactive sessions.
}
\examples{
\dontrun{
  check_and_install_cmdstanr()
}
}
