% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_if_any_NA}
\alias{check_if_any_NA}
\title{Check if NA}
\usage{
check_if_any_NA(.data, columns)
}
\arguments{
\item{.data}{A tibble including a gene name column | sample name column | read counts column | covariates column}

\item{.sample}{A column name}

\item{.cell_type}{A column name}

\item{.count}{A column name}

\item{.significance}{A column name}

\item{.do_check}{A column name}

\item{formula_columns}{A symbol vector}
}
\description{
Check if NA
}
