% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts_obj}
\alias{counts_obj}
\title{counts_obj}
\format{
A tidy data frame.
}
\usage{
data(counts_obj)
}
\description{
Example data set containing cell counts per cell cluster
}
\keyword{datasets}
