% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{do_inference}
\alias{do_inference}
\title{do_inference}
\usage{
do_inference(
  .data,
  formula,
  approximate_posterior_inference = F,
  approximate_posterior_analysis = F,
  C,
  X,
  cores,
  additional_parameters_to_save,
  to_exclude = tibble(N = integer(), M = integer()),
  truncation_compensation = 1,
  save_generated_quantities = F,
  inits_fx = "random",
  prior_from_discovery = tibble(.variable = character(), mean = numeric(), sd =
    numeric()),
  pass_fit = F,
  tol_rel_obj = 0.01,
  write_on_disk = F,
  seed,
  output_samples = output_samples,
  chains = 4
)
}
\arguments{
\item{formula}{A formula}

\item{approximate_posterior_inference}{A boolean}

\item{approximate_posterior_analysis}{A boolean}

\item{C}{An integer}

\item{X}{A tibble}

\item{cores}{An integer}

\item{additional_parameters_to_save}{A character vector}

\item{to_exclude}{A boolean}

\item{truncation_compensation}{A real}

\item{save_generated_quantities}{A boolean}

\item{inits_fx}{A function}

\item{prior_from_discovery}{A tibble}

\item{pass_fit}{A fit}

\item{tol_rel_obj}{A real}

\item{write_on_disk}{A boolean}

\item{seed}{an integer}

\item{my_df}{A tibble including a cell_type name column | sample name column | read counts column | covariates column}

\item{.sample}{A column name as symbol}

\item{.cell_type}{A column name as symbol}

\item{.count}{A column name as symbol}

\item{.significance}{A column name as symbol}

\item{.do_check}{A column name as symbol}

\item{lambda_mu_mu}{A real}

\item{exposure_rate_multiplier}{A real}

\item{intercept_shift_scale}{A real}

\item{adj_prob_theshold}{A real}

\item{how_many_posterior_draws}{A real number of posterior draws needed}
}
\value{
A tibble with additional columns
}
\description{
This function calls the stan model.
}
