% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fit_to_counts_rng}
\alias{fit_to_counts_rng}
\title{fit_to_counts_rng}
\usage{
fit_to_counts_rng(fit, adj_prob_theshold)
}
\arguments{
\item{fit}{A fit object}

\item{adj_prob_theshold}{fit real}
}
\description{
fit_to_counts_rng
}
