% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_output_samples}
\alias{get_output_samples}
\title{Get Output Samples from a Stan Fit Object}
\usage{
get_output_samples(fit)
}
\arguments{
\item{fit}{A \code{stanfit} object, which is the result of fitting a model via Stan.}
}
\value{
The number of output samples used in the Stan model.
Returns from MHC if available, otherwise from Variational inference.
}
\description{
This function retrieves the number of output samples from a Stan fit object,
supporting different methods (MHC and Variational) based on the available data within the object.
}
\examples{
# Assuming 'fit' is a stanfit object obtained from running a Stan model
samples_count = get_output_samples(fit)

}
