% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glm_dirichlet_multinomial}
\alias{glm_dirichlet_multinomial}
\title{glm_dirichlet_multinomial}
\format{
A text file containing stan code for the Dirichlet model.
}
\usage{
data(glm_dirichlet_multinomial)
}
\description{
glm_dirichlet_multinomial
}
\keyword{datasets}
