% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glm_dirichlet_multinomial_generate_quantities}
\alias{glm_dirichlet_multinomial_generate_quantities}
\title{glm_dirichlet_multinomial_generate_quantities}
\format{
A text file containing stan code for the Dirichlet model.
}
\usage{
data(glm_dirichlet_multinomial_generate_quantities)
}
\description{
This object is mostly for internal use and comparative purposes, if the dirichlet_multinomial is chosen as noise model.
}
\keyword{internal}
