% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glm_dirichlet_multinomial_imputation}
\alias{glm_dirichlet_multinomial_imputation}
\title{glm_dirichlet_multinomial_imputation}
\format{
A text file containing stan code for the Dirichlet model.
}
\usage{
data(glm_dirichlet_multinomial_imputation)
}
\description{
glm_dirichlet_multinomial_imputation
}
\keyword{datasets}
