% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glm_multi_beta_binomial_generate}
\alias{glm_multi_beta_binomial_generate}
\title{glm_multi_beta_binomial}
\format{
A text file containing stan code for the Dirichlet model.

A text file containing stan code for the Beta only model.
}
\usage{
data(glm_multi_beta_binomial)

data(glm_multi_beta_binomial_generate)
}
\description{
This object is for cmdstanr

This object is for cmdstanr
}
\keyword{internal}
