% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{harmonise_factor_levels}
\alias{harmonise_factor_levels}
\title{harmonise_factor_levels}
\usage{
harmonise_factor_levels(dataframe_query, dataframe_reference)
}
\arguments{
\item{new_data}{A data frame containing potential factor variables}

\item{old_data}{A data frame containing the reference factor levels}
}
\value{
A data frame with the same dimensions as \code{new_data}
}
\description{
A helper function to make sure that factor levels in new data match the old data
}
