% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_imputation}
\alias{model_imputation}
\title{model_imputation}
\format{
A text file containing stan code for the Dirichlet model.
}
\usage{
data(model_imputation)
}
\description{
This object is mostly for internal use and comparative purposes, if the dirichlet_multinomial is chosen as noise model.
}
\keyword{internal}
