% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mutate_from_expr_list}
\alias{mutate_from_expr_list}
\title{Mutate Data Frame Based on Expression List}
\usage{
mutate_from_expr_list(x, formula_expr, ignore_errors = TRUE)
}
\arguments{
\item{x}{A data frame to be mutated.}

\item{formula_expr}{A named list of formula expressions used for mutation.}

\item{ignore_errors}{Logical flag indicating whether to ignore errors during mutation.}
}
\value{
A mutated data frame with added or modified columns based on \code{formula_expr}.
}
\description{
\code{mutate_from_expr_list} takes a data frame and a list of formula expressions,
and mutates the data frame based on these expressions. It allows for ignoring
errors during the mutation process.
}
\details{
The function performs various checks and transformations on the formula expressions,
ensuring that the specified transformations are valid and can be applied to the data frame.
It supports advanced features like handling special characters in column names and intelligent
parsing of formulas.
}
