% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{no_significance_df}
\alias{no_significance_df}
\title{no_significance_df}
\format{
A tibble with the following columns:
\itemize{
\item \strong{sample}: Character. Identifier for each sample.
\item \strong{condition}: Character. Experimental condition or group (e.g., "X" or "Y").
\item \strong{cell_group}: Character. Cell group or cell type (e.g., "A", "B").
\item \strong{count}: Numeric. Count of cells in the given sample, condition, and cell group.
}
}
\usage{
data(no_significance_df)
}
\value{
A tibble with 34 rows and 4 columns: \code{sample}, \code{condition}, \code{cell_group}, and \code{count}.
}
\description{
A small example dataset containing cell counts across samples, conditions, and cell groups. This dataset is used to demonstrate the use of \code{sccomp} functions in scenarios where there is no significant difference in cell composition between conditions.
}
\examples{
data(no_significance_df)
head(no_significance_df)

}
\keyword{datasets}
