% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oligo_breast_estimate}
\alias{oligo_breast_estimate}
\title{oligo_breast_estimate}
\format{
A sccomp_glm result.
}
\usage{
data(oligo_breast_estimate)
}
\description{
oligo_breast_estimate
}
\keyword{datasets}
