% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_credible_intervals}
\alias{plot_credible_intervals}
\title{plot_credible interval for theoretical data distributions}
\usage{
plot_credible_intervals(.data)
}
\arguments{
\item{.data}{The tibble returned by sccomp_glm}
}
\value{
A tibble with an additional \code{plot} column
}
\description{
Plot the data along the theoretical data distribution.
}
