% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_OLD_framework.R
\name{plot_summary}
\alias{plot_summary}
\title{plot_summary}
\usage{
plot_summary(.data, significance_threshold = 0.025)
}
\arguments{
\item{.data}{A tibble including a cell_group name column | sample name column | read counts column | factor columns | Pvalue column | a significance column}

\item{significance_threshold}{A real. FDR threshold for labelling significant cell-groups.}
}
\value{
A \code{ggplot}
}
\description{
This function plots a summary of the results of the model.
}
\examples{

data("counts_obj")

estimate =
  sccomp_estimate(
  counts_obj ,
   ~ type, ~1, sample, cell_group, count,
    approximate_posterior_inference = "all",
    cores = 1
  )

# estimate |> plot_summary()

}
