% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_method.R
\name{print.sccomp_tbl}
\alias{print.sccomp_tbl}
\title{Print method for sccomp objects}
\usage{
\method{print}{sccomp_tbl}(x, ...)
}
\arguments{
\item{x}{A sccomp object}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the input object
}
\description{
Print method for sccomp objects
}
\examples{
# See sccomp_estimate() for examples
}
