% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{replicate_data}
\alias{replicate_data}
\title{replicate_data}
\usage{
replicate_data(.data, number_of_draws = 1, seed = 42)
}
\arguments{
\item{.data}{A tibble including a cell_type name column | sample name column | read counts column | covariate columns | Pvalue column | a significance column}

\item{number_of_draws}{An integer. How may copies of the data you want to draw from the model joint posterior distribution.}

\item{seed}{An integer. Used for development and testing purposes}
}
\value{
A nested tibble \code{tbl} with cell_group-wise statistics
}
\description{
This function replicates counts from a real-world dataset.
}
\examples{

data("counts_obj")

estimate =
  sccomp_glm(
  counts_obj ,
   ~ type,  sample, cell_group, count,
    approximate_posterior_inference = TRUE,
    check_outliers = FALSE
  )

}
