% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_obj}
\alias{seurat_obj}
\title{seurat_obj}
\format{
A Seurat object
}
\usage{
data(seurat_obj)
}
\description{
Example Seurat data set. Seurat data objects can be directly used with sccomp_glm function.
}
\keyword{datasets}
