% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{DEPRECATED: simulate_data}
\usage{
simulate_data(
  .data,
  .estimate_object,
  formula_composition,
  formula_variability = NULL,
  .sample = NULL,
  .cell_group = NULL,
  .coefficients = NULL,
  variability_multiplier = 1,
  number_of_draws = 1,
  mcmc_seed = sample_seed(),
  cores = detectCores(),
  sig_figs = 9,
  cache_stan_model = sccomp_stan_models_cache_dir
)
}
\arguments{
\item{.data}{A tibble including a cell_group name column | sample name column | read counts column | factor columns | Pvalue column | a significance column}

\item{.estimate_object}{The result of sccomp_estimate execution. This is used for sampling from real-data properties.}

\item{formula_composition}{A formula. The formula describing the model for differential abundance, for example ~treatment}

\item{formula_variability}{A formula. The formula describing the model for differential variability, for example ~treatment}

\item{.sample}{A column name as symbol. The sample identifier}

\item{.cell_group}{A column name as symbol. The cell_group identifier}

\item{.coefficients}{(Deprecated) The column names for coefficients in new_data, for example, c(b_0, b_1). Use the 'coefficients' parameter in sccomp_simulate() instead.}

\item{variability_multiplier}{A real scalar. This can be used for artificially increasing the variability of the simulation for benchmarking purposes.}

\item{number_of_draws}{An integer. How may copies of the data you want to draw from the model joint posterior distribution.}

\item{mcmc_seed}{An integer. Used for Markov-chain Monte Carlo reproducibility. By default a random number is sampled from 1 to 999999. This itself can be controlled by set.seed()}

\item{cores}{Integer, the number of cores to be used for parallel calculations.}

\item{sig_figs}{Number of significant figures to use for Stan model output. Default is 9.}

\item{cache_stan_model}{A character string specifying the cache directory for compiled Stan models.
The sccomp version will be automatically appended to ensure version isolation.
Default is \code{sccomp_stan_models_cache_dir} which points to \verb{~/.sccomp_models}.}
}
\description{
This function is DEPRECATED. Please use \code{\link{sccomp_simulate}} instead.
}
