% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{test_contrasts}
\alias{test_contrasts}
\title{test_contrasts}
\usage{
test_contrasts(
  .data,
  contrasts = NULL,
  percent_false_positive = 5,
  test_composition_above_logit_fold_change = 0.2
)
}
\arguments{
\item{.data}{A tibble. The result of sccomp_glm.}

\item{contrasts}{A vector of character strings. For example if your formula is \code{~ 0 + treatment} and the factor treatment has values \code{yes} and \code{no}, your contrast could be "constrasts = c(treatmentyes - treatmentno)".}

\item{percent_false_positive}{A real between 0 and 100 non included. This used to identify outliers with a specific false positive rate.}

\item{test_composition_above_logit_fold_change}{A positive integer. It is the effect threshold used for the hypothesis test. A value of 0.2 correspond to a change in cell proportion of 10\% for a cell type with baseline proportion of 50\%. That is, a cell type goes from 45\% to 50\%. When the baseline proportion is closer to 0 or 1 this effect thrshold has consistent value in the logit uncontrained scale.}
}
\value{
A nested tibble \code{tbl} with cell_group-wise statistics
}
\description{
This function test ocntrasts from a sccomp result.
}
\examples{

data("counts_obj")

  estimates =
  sccomp_glm(
  counts_obj ,
   ~ 0 + type, ~1,  sample, cell_group, count,
    check_outliers = FALSE,
    cores = 1
  ) |>

  test_contrasts("typecancer - typebenign")

}
