% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vb_iterative}
\alias{vb_iterative}
\title{vb_iterative}
\usage{
vb_iterative(
  model,
  output_samples,
  iter,
  tol_rel_obj,
  additional_parameters_to_save = c(),
  data,
  seed,
  ...
)
}
\arguments{
\item{model}{A Stan model}

\item{output_samples}{An integer of how many samples from posteriors}

\item{iter}{An integer of how many max iterations}

\item{tol_rel_obj}{A real}

\item{additional_parameters_to_save}{A character vector}

\item{...}{List of paramaters for vb function of Stan}
}
\value{
A Stan fit object
}
\description{
Runs iteratively variational bayes until it suceeds
}
