% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{clear_stan_model_cache}
\alias{clear_stan_model_cache}
\title{Clear Stan Model Cache}
\usage{
clear_stan_model_cache(cache_dir = sccomp_stan_models_cache_dir)
}
\arguments{
\item{cache_dir}{A character string representing the path of the cache directory to delete. Defaults to \code{sccomp_stan_models_cache_dir}.}
}
\description{
This function attempts to delete the Stan model cache directory and its contents.
If the cache directory does not exist, it prints a message indicating this.
}
\examples{
\dontrun{
  clear_stan_model_cache("path/to/cache_dir")
}
}
