% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glm_multi_beta}
\alias{glm_multi_beta}
\title{glm_multi_beta}
\format{
A text file containing stan code for the Beta only model.
}
\usage{
data(glm_multi_beta)
}
\description{
glm_multi_beta
}
\keyword{datasets}
