% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load_model}
\alias{load_model}
\title{Load, Compile, and Cache a Stan Model}
\usage{
load_model(name, cache_dir = sccomp_stan_models_cache_dir)
}
\arguments{
\item{name}{A character string representing the name of the Stan model.}

\item{cache_dir}{A character string representing the path to the cache directory.}
}
\value{
A compiled Stan model object.
}
\description{
This function attempts to load a precompiled Stan model using the \code{instantiate} package.
If the model is not found, it will locate the Stan model file within the \code{sccomp} package,
compile it using \code{cmdstanr}, and save the compiled model to the cache directory.
}
\examples{
\dontrun{
  model <- load_model("glm_multi_beta_binomial_", "~/cache")
}
}
