% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_multi_beta.R
\name{multi_beta_glm}
\alias{multi_beta_glm}
\title{multi_beta_glm main}
\usage{
multi_beta_glm(
  .data,
  formula = ~1,
  .sample,
  check_outliers = FALSE,
  approximate_posterior_inference = T,
  cores = detect_cores(),
  seed = sample(1:99999, size = 1)
)
}
\arguments{
\item{.data}{A tibble including a cell_type name column | sample name column | read counts column | covariate columns | Pvaue column | a significance column}

\item{formula}{A formula. The sample formula used to perform the differential cell_type abundance analysis}

\item{.sample}{A column name as symbol. The sample identifier}

\item{approximate_posterior_inference}{A boolean. Whether the inference of the joint posterior distribution should be approximated with variational Bayes. It confers execution time advantage.}

\item{cores}{An integer. How many cored to be used with parallel calculations.}

\item{seed}{An integer. Used for development and testing purposes}

\item{.cell_type}{A column name as symbol. The cell_type identifier}

\item{.count}{A column name as symbol. The cell_type abundance (read count)}
}
\value{
A nested tibble \code{tbl} with cell_type-wise information: \verb{sample wise data} | plot | \verb{ppc samples failed} | \verb{exposure deleterious outliers}
}
\description{
This function runs the data modelling and statistical test for the hypothesis that a cell_type includes outlier biological replicate.
}
