% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multipanel_theme}
\alias{multipanel_theme}
\title{multipanel_theme}
\format{
A ggplot theme
}
\usage{
data(multipanel_theme)
}
\description{
this object includes the ggplot, theme
}
\keyword{datasets}
