% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{remove_unwanted_variation}
\alias{remove_unwanted_variation}
\title{remove_unwanted_variation}
\usage{
remove_unwanted_variation(
  .data,
  formula_composition = ~1,
  formula_variability = NULL
)
}
\arguments{
\item{.data}{A tibble. The result of sccomp_glm.}

\item{formula_composition}{A formula. The formula describing the model for differential abundance, for example ~treatment. This formula can be a sub-formula of your estimated model; in this case all other covariate will be factored out.}

\item{formula_variability}{A formula. The formula describing the model for differential variability, for example ~treatment. In most cases, if differentially variability is of interest, the formula should only include the factor of interest as a large anount of data is needed to define variability depending to each covariates. This formula can be a sub-formula of your estimated model; in this case all other covariate will be factored out.}
}
\value{
A nested tibble \code{tbl} with cell_group-wise statistics
}
\description{
This function uses the model to remove unwanted variation from a dataset using the estimated of the model. For example if you fit your data with this formula \code{~ covariate_1 + covariate_2} and use this formula to remove unwanted variation \code{~ covariate_1}, the \code{covariate_2} will be factored out.
}
\examples{

data("counts_obj")

  estimates = sccomp_glm(
  counts_obj ,
   ~ type, ~1,  sample, cell_group, count,
    approximate_posterior_inference = "all",
    check_outliers = FALSE,
    cores = 1
  )

  remove_unwanted_variation(estimates)

}
