% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sccomp_glm}
\alias{sccomp_glm}
\title{sccomp_glm main}
\usage{
sccomp_glm(
  .data,
  formula = ~1,
  .sample,
  .cell_group,
  .count = NULL,
  formula_variability = ~1,
  prior_mean_variable_association = list(intercept = c(4.436925, 1.304049), slope =
    c(-0.73074903, 0.06532897), standard_deviation = c(5.06983, 8.549324)),
  percent_false_positive = 5,
  check_outliers = TRUE,
  approximate_posterior_inference = "outlier_detection",
  test_composition_above_logit_fold_change = 0.2,
  verbose = FALSE,
  noise_model = "multi_beta_binomial",
  variance_association = FALSE,
  exclude_priors = FALSE,
  cores = detectCores(),
  mcmc_seed = sample(1e+05, 1),
  max_sampling_iterations = 20000,
  pass_fit = TRUE
)

sccomp_glm(
  .data,
  formula = ~1,
  .sample,
  .cell_group,
  .count = NULL,
  formula_variability = ~1,
  prior_mean_variable_association = list(intercept = c(4.436925, 1.304049), slope =
    c(-0.73074903, 0.06532897), standard_deviation = c(5.06983, 8.549324)),
  percent_false_positive = 5,
  check_outliers = TRUE,
  approximate_posterior_inference = "outlier_detection",
  test_composition_above_logit_fold_change = 0.2,
  verbose = FALSE,
  noise_model = "multi_beta_binomial",
  variance_association = FALSE,
  exclude_priors = FALSE,
  cores = detectCores(),
  mcmc_seed = sample(1e+05, 1),
  max_sampling_iterations = 20000,
  pass_fit = TRUE
)
}
\arguments{
\item{.data}{A tibble including a cell_type name column | sample name column | read counts column | covariate columns | Pvalue column | a significance column}

\item{formula}{A formula. The sample formula used to perform the differential cell_type abundance analysis}

\item{.sample}{A column name as symbol. The sample identifier}

\item{.cell_group}{A column name as symbol. The cell_type identifier}

\item{.count}{A column name as symbol. The cell_type abundance (read count). Used only for data frame count output.}

\item{prior_mean_variable_association}{A list of the form list(intercept = c(4.436925, 1.304049), slope = c(-0.73074903,  0.06532897), standard_deviation = c(5.06983, 8.549324)). Where for each parameter, we specify mean and standard deviation. This is used to incorporate prior knowledge about the mean/variability association of cell-type proportions.}

\item{percent_false_positive}{A real between 0 and 100. It is the aimed percent of cell types being a false positive. For example, percent_false_positive_genes = 1 provide 1 percent of the calls for significant changes that are actually not significant.}

\item{check_outliers}{A boolean. Whether to check for outliers before the fit.}

\item{approximate_posterior_inference}{A boolean. Whether the inference of the joint posterior distribution should be approximated with variational Bayes. It confers execution time advantage.}

\item{test_composition_above_logit_fold_change}{A positive integer. It is the effect threshold used for the hypothesis test. A value of 0.2 correspond to a change in cell proportion of 10\% for a cell type with baseline proportion of 50\%. That is, a cell type goes from 45\% to 50\%. When the baseline proportion is closer to 0 or 1 this effect thrshold has consistent value in the logit uncontrained scale.}

\item{verbose}{A boolean. Prints progression.}

\item{noise_model}{A character string. The two noise models available are multi_beta_binomial (default) and dirichlet_multinomial.}

\item{variance_association}{A boolean. Whether the variance should depend on the factor of interest.}

\item{cores}{An integer. How many cored to be used with parallel calculations.}

\item{mcmc_seed}{An integer. Used for Markov-chain Monte Carlo reproducibility. By default a random number is sampled from 1 to 999999. This itself can be controlled by set.seed()}
}
\value{
A nested tibble \code{tbl} with cell_group-wise statistics

A nested tibble \code{tbl} with cell_group-wise statistics
}
\description{
This function runs the data modelling and statistical test for the hypothesis that a cell_type includes outlier biological replicate.

This function runs the data modelling and statistical test for the hypothesis that a cell_type includes outlier biological replicate.
}
\examples{

data("counts_obj")

estimate =
  sccomp_glm(
  counts_obj ,
   ~ type,  sample, cell_group, count,
    approximate_posterior_inference = "all",
    check_outliers = FALSE,
    cores = 1
  )


data("counts_obj")

estimate =
  sccomp_glm(
  counts_obj ,
   ~ type,  sample, cell_group, count,
    approximate_posterior_inference = "all",
    check_outliers = FALSE,
    cores = 1
  )

}
