% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sccomp_glm}
\alias{sccomp_glm}
\title{sccomp_glm main}
\usage{
sccomp_glm(
  .data,
  formula = ~1,
  .sample,
  .cell_group,
  .count = NULL,
  prior_mean_variable_association = list(intercept = c(4.436925, 1.304049), slope =
    c(-0.73074903, 0.06532897), standard_deviation = c(0.4527292, 0.3318759)),
  percent_false_positive = 5,
  check_outliers = TRUE,
  approximate_posterior_inference = FALSE,
  verbose = FALSE,
  noise_model = "multi_beta_binomial",
  variance_association = FALSE,
  cores = detectCores(),
  seed = 42
)
}
\arguments{
\item{.data}{A tibble including a cell_type name column | sample name column | read counts column | covariate columns | Pvalue column | a significance column}

\item{formula}{A formula. The sample formula used to perform the differential cell_type abundance analysis}

\item{.sample}{A column name as symbol. The sample identifier}

\item{.cell_group}{A column name as symbol. The cell_type identifier}

\item{.count}{A column name as symbol. The cell_type abundance (read count). Used only for data frame count output.}

\item{percent_false_positive}{A real between 0 and 100. It is the aimed percent of cell types being a false positive. For example, percent_false_positive_genes = 1 provide 1 percent of the calls for significant changes that are actually not significant.}

\item{check_outliers}{A boolean. Whether to check for outliers before the fit.}

\item{approximate_posterior_inference}{A boolean. Whether the inference of the joint posterior distribution should be approximated with variational Bayes. It confers execution time advantage.}

\item{verbose}{A boolean. Prints progression.}

\item{noise_model}{A character string. The two noise models available are multi_beta_binomial (default) and dirichlet_multinomial.}

\item{cores}{An integer. How many cored to be used with parallel calculations.}

\item{seed}{An integer. Used for development and testing purposes}
}
\value{
A nested tibble \code{tbl} with cell_group-wise statistics
}
\description{
This function runs the data modelling and statistical test for the hypothesis that a cell_type includes outlier biological replicate.
}
\examples{

data("counts_obj")

estimate =
  sccomp_glm(
  counts_obj ,
   ~ type,  sample, cell_group, count,
    approximate_posterior_inference = FALSE
  )

}
