% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sccomp_glm}
\alias{sccomp_glm}
\title{sccomp_glm main}
\usage{
sccomp_glm(
  .data,
  formula_composition = ~1,
  formula_variability = ~1,
  .sample,
  .cell_group,
  .count = NULL,
  contrasts = NULL,
  prior_mean_variable_association = list(intercept = c(5, 2), slope = c(0, 0.6),
    standard_deviation = c(20, 40)),
  check_outliers = TRUE,
  bimodal_mean_variability_association = FALSE,
  enable_loo = FALSE,
  cores = detectCores(),
  percent_false_positive = 5,
  approximate_posterior_inference = "none",
  test_composition_above_logit_fold_change = 0.2,
  .sample_cell_group_pairs_to_exclude = NULL,
  verbose = FALSE,
  noise_model = "multi_beta_binomial",
  exclude_priors = FALSE,
  use_data = TRUE,
  mcmc_seed = sample(1e+05, 1),
  max_sampling_iterations = 20000,
  pass_fit = TRUE
)
}
\arguments{
\item{.data}{A tibble including a cell_group name column | sample name column | read counts column (optional depending on the input class) | factor columns.}

\item{formula_composition}{A formula. The formula describing the model for differential abundance, for example ~treatment.}

\item{formula_variability}{A formula. The formula describing the model for differential variability, for example ~treatment. In most cases, if differentially variability is of interest, the formula should only include the factor of interest as a large anount of data is needed to define variability depending to each factors.}

\item{.sample}{A column name as symbol. The sample identifier}

\item{.cell_group}{A column name as symbol. The cell_group identifier}

\item{.count}{A column name as symbol. The cell_group abundance (read count). Used only for data frame count output. The variable in this column should be of class integer.}

\item{contrasts}{A vector of character strings. For example if your formula is \code{~ 0 + treatment} and the factor treatment has values \code{yes} and \code{no}, your contrast could be constrasts = c("treatmentyes - treatmentno").}

\item{prior_mean_variable_association}{A list of the form list(intercept = c(5, 2), slope = c(0,  0.6), standard_deviation = c(20, 40)). Where for intercept and slope parameters, we specify mean and standard deviation, while for standard deviation, we specify shape and rate. This is used to incorporate prior knowledge about the mean/variability association of cell-type proportions.}

\item{check_outliers}{A boolean. Whether to check for outliers before the fit.}

\item{bimodal_mean_variability_association}{A boolean. Whether to model the mean-variability as bimodal, as often needed in the case of single-cell RNA sequencing data, and not usually for CyTOF and microbiome data. The plot summary_plot()$credible_intervals_2D can be used to assess whether the bimodality should be modelled.}

\item{enable_loo}{A boolean. Enable model comparison by the R package LOO. This is helpful when you want to compare the fit between two models, for example, analogously to ANOVA, between a one factor model versus a interceot-only model.}

\item{cores}{An integer. How many cored to be used with parallel calculations.}

\item{percent_false_positive}{A real between 0 and 100 non included. This used to identify outliers with a specific false positive rate.}

\item{approximate_posterior_inference}{A boolean. Whether the inference of the joint posterior distribution should be approximated with variational Bayes. It confers execution time advantage.}

\item{test_composition_above_logit_fold_change}{A positive integer. It is the effect threshold used for the hypothesis test. A value of 0.2 correspond to a change in cell proportion of 10\% for a cell type with baseline proportion of 50\%. That is, a cell type goes from 45\% to 50\%. When the baseline proportion is closer to 0 or 1 this effect thrshold has consistent value in the logit uncontrained scale.}

\item{.sample_cell_group_pairs_to_exclude}{A column name that includes a boolean variable for the sample/cell-group pairs to be ignored in the fit. This argument is for pro-users.}

\item{verbose}{A boolean. Prints progression.}

\item{noise_model}{A character string. The two noise models available are multi_beta_binomial (default) and dirichlet_multinomial.}

\item{exclude_priors}{A boolean. Whether to run a prior-free model, for benchmarking purposes.}

\item{use_data}{A booelan. Whether to sun the model data free. This can be used for prior predictive check.}

\item{mcmc_seed}{An integer. Used for Markov-chain Monte Carlo reproducibility. By default a random number is sampled from 1 to 999999. This itself can be controlled by set.seed()}

\item{max_sampling_iterations}{An integer. This limit the maximum number of iterations in case a large dataset is used, for limiting the computation time.}

\item{pass_fit}{A boolean. Whether to pass the Stan fit as attribute in the output. Because the Stan fit can be very large, setting this to FALSE can be used to lower the memory imprint to save the output.}
}
\value{
A nested tibble \code{tbl}, with the following columns
\itemize{
\item cell_group - column including the cell groups being tested
\item parameter - The parameter being estimated, from the design matrix dscribed with the input formula_composition and formula_variability
\item factor - The factor in the formula corresponding to the covariate, if exists (e.g. it does not exist in case og Intercept or contrasts, which usually are combination of parameters)

\item c_lower - lower (2.5\%) quantile of the posterior distribution for a composition (c) parameter.
\item c_effect - mean of the posterior distribution for a composition (c) parameter.
\item c_upper - upper (97.5\%) quantile of the posterior distribution fo a composition (c)  parameter.
\item c_pH0 - Probability of the null hypothesis (no difference) for  a composition (c). This is not a p-value.
\item c_FDR - False-discovery rate of the null hypothesis (no difference) for  a composition (c).
\item c_n_eff - Effective sample size - the number of independent draws in the sample, the higher the better (mc-stan.org/docs/2_25/cmdstan-guide/stansummary.html).
\item c_R_k_hat - R statistic, a measure of chain equilibrium, should be within 0.05 of 1.0 (mc-stan.org/docs/2_25/cmdstan-guide/stansummary.html).

\item v_lower - Lower (2.5\%) quantile of the posterior distribution for a variability (v) parameter
\item v_effect - Mean of the posterior distribution for a variability (v) parameter
\item v_upper - Upper (97.5\%) quantile of the posterior distribution for a variability (v) parameter
\item v_pH0 - Probability of the null hypothesis (no difference) for a variability (v). This is not a p-value.
\item v_FDR - False-discovery rate of the null hypothesis (no difference), for a variability (v).
\item v_n_eff - Effective sample size for a variability (v) parameter - the number of independent draws in the sample, the higher the better (mc-stan.org/docs/2_25/cmdstan-guide/stansummary.html).
\item v_R_k_hat - R statistic for a variability (v) parameter, a measure of chain equilibrium, should be within 0.05 of 1.0 (mc-stan.org/docs/2_25/cmdstan-guide/stansummary.html).

\item count_data Nested input count data.

}
}
\description{
The function for linear modelling takes as input a table of cell counts with three columns containing a cell-group identifier, sample identifier, integer count and the factors (continuous or discrete). The user can define a linear model with an input R formula, where the first factor is the factor of interest. Alternatively, sccomp accepts single-cell data containers (Seurat, SingleCellExperiment44, cell metadata or group-size). In this case, sccomp derives the count data from cell metadata.
}
\examples{

data("counts_obj")

estimate =
  sccomp_glm(
  counts_obj ,
   ~ type,
   ~1,
   sample,
   cell_group,
   count,
    check_outliers = FALSE,
    cores = 1
  )

}
