% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sccomp_glm}
\alias{sccomp_glm}
\title{sccomp_glm main}
\usage{
sccomp_glm(
  .data,
  formula,
  .sample,
  .cell_group,
  .count = NULL,
  check_outliers = TRUE,
  approximate_posterior_inference = FALSE,
  verbose = FALSE,
  noise_model = "multi_beta_binomial",
  cores = detectCores(),
  seed = 42
)
}
\arguments{
\item{.data}{A tibble including a cell_type name column | sample name column | read counts column | covariate columns | Pvaue column | a significance column}

\item{formula}{A formula. The sample formula used to perform the differential cell_type abundance analysis}

\item{.sample}{A column name as symbol. The sample identifier}

\item{.cell_group}{A column name as symbol. The cell_type identifier}

\item{.count}{A column name as symbol. The cell_type abundance (read count)}

\item{approximate_posterior_inference}{A boolean. Whether the inference of the joint posterior distribution should be approximated with variational Bayes. It confers execution time advantage.}

\item{cores}{An integer. How many cored to be used with parallel calculations.}

\item{seed}{An integer. Used for development and testing purposes}
}
\value{
A nested tibble \code{tbl} with cell_group-wise statistics
}
\description{
This function runs the data modelling and statistical test for the hypothesis that a cell_type includes outlier biological replicate.
}
