% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sccomp_predict}
\alias{sccomp_predict}
\title{sccomp_predict}
\usage{
sccomp_predict(
  fit,
  formula_composition = NULL,
  new_data = NULL,
  number_of_draws = 500,
  mcmc_seed = sample(1e+05, 1)
)
}
\arguments{
\item{fit}{The result of sccomp_glm.}

\item{formula_composition}{A formula. The formula describing the model for differential abundance, for example ~treatment. This formula can be a sub-formula of your estimated model; in this case all other factor will be factored out.}

\item{new_data}{A sample-wise data frame including the column that represent the factors in your formula. If you want to predict proportions for 10 samples, there should be 10 rows. T}

\item{number_of_draws}{An integer. How may copies of the data you want to draw from the model joint posterior distribution.}

\item{mcmc_seed}{An integer. Used for Markov-chain Monte Carlo reproducibility. By default a random number is sampled from 1 to 999999. This itself can be controlled by set.seed()}
}
\value{
A nested tibble \code{tbl} with cell_group-wise statistics
}
\description{
This function replicates counts from a real-world dataset.
}
\examples{

data("counts_obj")

if(.Platform$OS.type == "unix")
  sccomp_glm(
  counts_obj ,
   ~ type, ~1,  sample, cell_group, count,
    approximate_posterior_inference = "all",
    check_outliers = FALSE,
    cores = 1
  ) |>

  sccomp_predict()

}
