% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_obj}
\alias{sce_obj}
\title{sce_obj}
\format{
A SingeCellExperiment object. SingeCellExperiment data objects can be directly used with sccomp_glm function.
}
\usage{
data(sce_obj)
}
\description{

}
\keyword{datasets}
