% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_obj}
\alias{sce_obj}
\title{Example data set}
\format{
An object of class \code{SingleCellExperiment} with 1 rows and 106297 columns.
}
\usage{
sce_obj
}
\description{
Example data set
}
\keyword{datasets}
