% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seurat_obj}
\alias{seurat_obj}
\title{Example data set}
\format{
An object of class \code{Seurat} with 1 rows and 106297 columns.
}
\usage{
seurat_obj
}
\description{
Example data set
}
\keyword{datasets}
