% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{simulate_data}
\alias{simulate_data}
\title{simulate_data}
\usage{
simulate_data(
  .data,
  .estimate_object,
  formula,
  .sample = NULL,
  .cell_group = NULL,
  .coefficients = NULL,
  number_of_draws = 1,
  mcmc_seed = sample(1e+05, 1)
)
}
\arguments{
\item{.data}{A tibble including a cell_type name column | sample name column | read counts column | covariate columns | Pvalue column | a significance column}

\item{.estimate_object}{The result of sccomp_glm execution. This is used for sampling from real-data properties.}

\item{formula}{A formula. The sample formula used to perform the differential cell_type abundance analysis}

\item{.sample}{A column name as symbol. The sample identifier}

\item{.cell_group}{A column name as symbol. The cell_type identifier}

\item{.coefficients}{Th column names for coefficients, for example, c(b_0, b_1)}

\item{number_of_draws}{An integer. How may copies of the data you want to draw from the model joint posterior distribution.}

\item{mcmc_seed}{An integer. Used for Markov-chain Monte Carlo reproducibility. By default a random number is sampled from 1 to 999999. This itself can be controlled by set.seed()}
}
\value{
A nested tibble \code{tbl} with cell_group-wise statistics
}
\description{
This function simulates counts from a linear model.
}
\examples{

data("counts_obj")
library(dplyr)

estimate =
  sccomp_glm(
  counts_obj ,
   ~ type,  sample, cell_group, count,
    approximate_posterior_inference = "all",
    check_outliers = FALSE,
    cores = 1
  )

# Set coefficients for cell_types. In this case all coefficients are 0 for simplicity.
counts_obj = counts_obj |> mutate(b_0 = 1, b_1 = 1)

# Simulate data
simulate_data(counts_obj, estimate, ~type, sample, cell_group, c(b_0, b_1))

}
